requirements.txt
torchvision==0.5.0
efficientnet_pytorch==0.5.1
albumentations==0.4.3
torch==1.7.1+cpu
Pillow==9.1.0
streamlit==1.9.0
